Fix audit findings: compile errors, type safety, a11y, code quality (#621-#643)
- Fix TS compile errors in FluidAudioDiarizer, OnnxWebTranslator, StreamingProcessor
- Add missing ElectronAPI type definitions for drag mode and shortcut IPC
- Fix KotobaWhisperEngine type incompatibility with base class
- Fix ESLint violations (unused import, unnecessary escape, stale directive)
- Remove OnnxWebSTTEngine stub class and registration
- Add listener cleanup in auto-updater and WebSocket audio dispose
- Replace silent .catch(() => {}) with console.warn logging (10 locations)
- Extract GlossarySettings component from TranslatorSettings (782→~430 lines)
- Extract cleanupResources helper in useAudioCapture (3 duplicates → 1)
- Extract shared audio constants (MIN_AUDIO_CHUNK_SAMPLES, SILENCE_THRESHOLD)
- Extract MDM_READ_TIMEOUT_MS constant
- Improve subtitle overlay a11y: touch targets, aria-labels, text overflow
- Add aria-expanded to collapsible settings sections
- Add loading/empty states to KeyboardShortcuts and CorrectionHistory
- Fix color contrast for helper text (#64748b → #94a3b8)
- Increase INTERIM_TEXT_COLOR opacity from 0.5 to 0.7 for WCAG compliance
- Add flex-wrap to language selector for narrow layouts
- Document worker-pool refCount safety (single-threaded JS)
