+++ src/engines/stt/SenseVoiceSherpaEngine.ts
@@ -253,7 +253,6 @@ export class SenseVoiceSherpaEngine implements STTEngine {
       const ws = createWriteStream(archivePath)
       let downloaded = 0
 
-      // eslint-disable-next-line no-constant-condition
       while (true) {
         const { done, value } = await reader.read()
         if (done) break
