+++ src/main/audio-constants.ts
@@ -1,0 +1,5 @@
+/** Minimum audio chunk size in samples (0.5s at 16kHz) */
+export const MIN_AUDIO_CHUNK_SAMPLES = 8000
+
+/** Amplitude threshold below which audio is considered silence */
+export const SILENCE_THRESHOLD = 0.001
