+++ src/main/audio-handlers.ts
@@ -2,15 +2,10 @@ import { ipcMain } from 'electron'
 import { sanitizeErrorMessage } from './error-utils'
 import { createLogger } from './logger'
 import { SAMPLE_RATE } from './constants'
+import { MIN_AUDIO_CHUNK_SAMPLES, SILENCE_THRESHOLD } from './audio-constants'
 import type { AppContext } from './app-context'
 
 const log = createLogger('audio')
-
-/** Minimum number of samples for a valid audio chunk (0.5s at 16kHz) */
-const MIN_AUDIO_CHUNK_SAMPLES = 8000
-
-/** Minimum amplitude to consider a chunk as non-silent */
-const SILENCE_THRESHOLD = 0.001
 
 /** Convert IPC audio data to Float32Array, returning null for silence */
 function toFloat32Array(audioData: unknown): Float32Array | null {
